+++ falco-misp-connector.py
@@ -47,7 +47,7 @@ sha256_dict = {}
 ip6_outbound_list = []
 file_list = []
 uri_list = []
-cidr_list = []
+cidr_outbound_list = []
 
 
 #################################################
@@ -55,7 +55,7 @@ cidr_list = []
 #################################################
 
 print("Contacting MISP Server: " + str(misp_server_url) )
-ip4_outbound_list, ip6_outbound_list, domain_list, file_list, sha256_dict, uri_list, cidr_list = fetchMISPIndicators(ip4_outbound_list, ip6_outbound_list, domain_list, file_list, sha256_dict, uri_list,cidr_list)
+ip4_outbound_list, ip6_outbound_list, domain_list, file_list, sha256_dict, uri_list, cidr_outbound_list = fetchMISPIndicators(ip4_outbound_list, ip6_outbound_list, domain_list, file_list, sha256_dict, uri_list,cidr_outbound_list)
 
 
 
@@ -63,7 +63,7 @@ ip4_outbound_list, ip6_outbound_list, domain_list, file_list, sha256_dict, uri_l
 # Convert arrays to a string ready for writing  #
 #################################################
 ip4_outbound_list_output_str =createYAMLArray(ip4_outbound_list) 
-cidr_list_output_str =createYAMLArray(cidr_list) 
+cidr_outbound_list_output_str =createYAMLArray(cidr_outbound_list) 
 
 #  The IPV6 addresses need speech marks around them
 ip6_outbound_list_output_str =createYAMLArray(ip6_outbound_list) 
@@ -76,18 +76,18 @@ ip6_outbound_list_output_str = ip6_outbound_list_output_str.replace(",", "\",\""
 # This makes it easier to pass Falco validation                   #
 ###################################################################
 # read: sample-falco-ipv4-rule.yaml
-with open("./sample-falco-ipv4-rule.yaml", 'r') as file:
+with open("./sample-falco-ipv4-outbound-rule.yaml", 'r') as file:
     ipv4_rule_content = file.read()
 ip4_outbound_list_output_str = ip4_outbound_list_output_str + "\n\n" + ipv4_rule_content
 
 # read: sample-falco-ipv6-rule.yaml
-with open("./sample-falco-ipv6-rule.yaml", 'r') as file:
+with open("./sample-falco-ipv6-outbound-rule.yaml", 'r') as file:
     ipv6_rule_content = file.read()
 ip6_outbound_list_output_str = ip6_outbound_list_output_str + "\n\n" + ipv6_rule_content
 
-with open("./sample-falco-cidr-rule.yaml", 'r') as file:
+with open("./sample-falco-cidr-outbound-rule.yaml", 'r') as file:
     cidr_rule_content = file.read()
-cidr_list_output_str = cidr_list_output_str + "\n\n" + cidr_rule_content
+cidr_outbound_list_output_str = cidr_outbound_list_output_str + "\n\n" + cidr_rule_content
 
 
 ###########################################################
@@ -98,18 +98,18 @@ if 'debugtest' in globals() and debugtest == True:
     print("Writing IPv4/IPv6 test valiation file")
     ip4_newline_list_output_str = createNLArray(ip4_outbound_list)
     ip6_newline_list_output_str = createNLArray(ip6_outbound_list)
-    cidr_newline_list_output_str = createNLArray(cidr_list)
+    cidr_newline_list_output_str = createNLArray(cidr_outbound_list)
     ip4_ip6_cidr_newline_list_output_str = ip4_newline_list_output_str +  ip6_newline_list_output_str + cidr_newline_list_output_str
     writeNewlineFile("tests/ip46.test", ip4_ip6_cidr_newline_list_output_str )
 
     print("Writing IPv4 Rules file to test folder")
-    writeFalcoRulesFileYaml("tests/ipv4-rules.yaml", falco_ipv4_list_name, ip4_outbound_list_output_str)
+    writeFalcoRulesFileYaml("tests/ipv4-rules.yaml", falco_ipv4_outbound_list_name, ip4_outbound_list_output_str)
 
     print("Writing IPv6 Rules file to test folder")
-    writeFalcoRulesFileYaml("tests/ipv6-rules.yaml", falco_ipv6_list_name, ip6_outbound_list_output_str)
+    writeFalcoRulesFileYaml("tests/ipv6-rules.yaml", falco_ipv6_outbound_list_name, ip6_outbound_list_output_str)
 
     print("Writing CIDR Rules file to test folder")
-    writeFalcoRulesFileYaml("tests/cidr-rules.yaml", falco_cidr_list_name, cidr_list_output_str)
+    writeFalcoRulesFileYaml("tests/cidr-rules.yaml", falco_cidr_outbound_list_name, cidr_outbound_list_output_str)
 
     print("Finished writing validation files - exiting")
     sys.exit(0)
@@ -118,19 +118,19 @@ if 'debugtest' in globals() and debugtest == True:
 #   Update the items in the Falco rules files for IP   #
 ########################################################
 if debugyaml == True: print("- IPv4 YAML:" + str(ip4_outbound_list_output_str))
-print("Writing out IP indicators to: " + falco_ipv4_rules_file)
-writeFalcoRulesFileYaml(falco_ipv4_rules_file, falco_ipv4_list_name, ip4_outbound_list_output_str)
+print("Writing out IP indicators to: " + falco_ipv4_outbound_rules_file)
+writeFalcoRulesFileYaml(falco_ipv4_outbound_rules_file, falco_ipv4_outbound_list_name, ip4_outbound_list_output_str)
 
 ########################################################
 #   Update the items in the Falco rules files for IP   #
 ########################################################
 if debugyaml == True: print("- IPv6 YAML:" + str(ip6_outbound_list_output_str))
-print("Writing out IP indicators to: " + falco_ipv6_rules_file)
-writeFalcoRulesFileYaml(falco_ipv6_rules_file, falco_ipv6_list_name, ip6_outbound_list_output_str)
+print("Writing out IP indicators to: " + falco_ipv6_outbound_rules_file)
+writeFalcoRulesFileYaml(falco_ipv6_outbound_rules_file, falco_ipv6_outbound_list_name, ip6_outbound_list_output_str)
 
 ########################################################
 #   Update the items in the Falco rules files for CIDR #
 ########################################################
-if debugyaml == True: print("- CIDR YAML:" + str(cidr_list_output_str))
-print("Writing out CIDR indicators to: " + falco_cidr_rules_file)
-writeFalcoRulesFileYaml(falco_cidr_rules_file, falco_cidr_list_name, cidr_list_output_str)
+if debugyaml == True: print("- CIDR YAML:" + str(cidr_outbound_list_output_str))
+print("Writing out CIDR indicators to: " + falco_cidr_outbound_rules_file)
+writeFalcoRulesFileYaml(falco_cidr_outbound_rules_file, falco_cidr_outbound_list_name, cidr_outbound_list_output_str)