+++ parse_config.py
@@ -4,23 +4,23 @@ import sys
 
 def parseConfig():
     # Check Parameters Exist
-    if not 'falco_ipv4_rules_file' in globals():
-        print("Configuration file parameter 'falco_ipv4_rules_file' does not seem to exist")
+    if not 'falco_ipv4_outbound_rules_file' in globals():
+        print("Configuration file parameter 'falco_ipv4_outbound_rules_file' does not seem to exist")
         sys.exit(0)
-    if not 'falco_ipv4_list_name' in globals():
-        print("Configuration file parameter 'falco_ipv4_list_name' does not seem to exist")
+    if not 'falco_ipv4_outbound_list_name' in globals():
+        print("Configuration file parameter 'falco_ipv4_outbound_list_name' does not seem to exist")
         sys.exit(0)
-    if not 'falco_ipv6_rules_file' in globals():
-        print("Configuration file parameter 'falco_ipv6_rules_file' does not seem to exist")
+    if not 'falco_ipv6_outbound_rules_file' in globals():
+        print("Configuration file parameter 'falco_ipv6_outbound_rules_file' does not seem to exist")
         sys.exit(0)
-    if not 'falco_ipv6_list_name' in globals():
-        print("Configuration file parameter 'falco_ipv6_list_name' does not seem to exist")
+    if not 'falco_ipv6_outbound_list_name' in globals():
+        print("Configuration file parameter 'falco_ipv6_outbound_list_name' does not seem to exist")
         sys.exit(0)
-    if not 'falco_cidr_rules_file' in globals():
-        print("Configuration file parameter 'falco_cidr_rules_file' does not seem to exist")
+    if not 'falco_cidr_outbound_rules_file' in globals():
+        print("Configuration file parameter 'falco_cidr_outbound_rules_file' does not seem to exist")
         sys.exit(0)
-    if not 'falco_cidr_list_name' in globals():
-        print("Configuration file parameter 'falco_cidr_list_name' does not seem to exist")
+    if not 'falco_cidr_outbound_list_name' in globals():
+        print("Configuration file parameter 'falco_cidr_outbound_list_name' does not seem to exist")
         sys.exit(0)
     if not 'debug' in globals():
         print("Configuration file parameter 'debug' does not seem to exist")
@@ -45,13 +45,13 @@ def parseConfig():
     
     
     # Check Parameters are valid.
-    if not re.match('^(.+)\/([^\/]+)$',falco_ipv4_rules_file):
+    if not re.match('^(.+)\/([^\/]+)$',falco_ipv4_outbound_rules_file):
         print("Configuration file parameter 'falco_ipv4_rules_file' does not seem to be a file.  Please check parameter and try again")
         sys.exit(0)
-    if not re.match('^(.+)\/([^\/]+)$',falco_ipv6_rules_file):
+    if not re.match('^(.+)\/([^\/]+)$',falco_ipv6_outbound_rules_file):
         print("Configuration file parameter 'falco_ipv4_rules_file' does not seem to be a file.  Please check parameter and try again")
         sys.exit(0)
-    if not re.match('^(.+)\/([^\/]+)$',falco_cidr_rules_file):
+    if not re.match('^(.+)\/([^\/]+)$',falco_cidr_outbound_rules_file):
         print("Configuration file parameter 'falco_cidr_rules_file' does not seem to be a file.  Please check parameter and try again")
         sys.exit(0)
     if not re.match('(?:[a-z0-9](?:[a-z0-9-]{0,61}[a-z0-9])?\.)+[a-z0-9][a-z0-9-]{0,61}[a-z0-9]',misp_server_url):